requirements.txt
openai-whisper
streamlit
argparse
ffmpeg
typing
warnings
os
translate
pysrt
time
setuptools
